move format functions to helpers file
